trading_bot/requirements.txt
# Trading Bot Dependencies
# Install with: pip install -r requirements.txt

# Alpaca Trading API
alpaca-trade-api>=3.0.0

# Data manipulation and analysis
pandas>=2.0.0
numpy>=1.24.0

# Technical analysis
ta>=0.10.2
pandas-ta>=0.3.14b

# Excel spreadsheet handling
openpyxl>=3.1.0

# Scheduling and time handling
schedule>=1.2.0
pytz>=2023.3

# HTTP requests (for API calls)
requests>=2.31.0

# Email notifications
# (uses built-in smtplib, no extra package needed)

# Logging (uses built-in logging module)

# Date utilities
python-dateutil>=2.8.2

# Environment variables
python-dotenv>=1.0.0
